chore: refactor member update event
