+++ src/events/memberEvents/memberUpdate.ts
@@ -6,9 +6,10 @@ import {
 } from "discord.js";
 import { getFixedT } from "i18next";
 
-import { defaultServer } from "../../config/database/defaultServer";
 import { BeccaLyria } from "../../interfaces/BeccaLyria";
-import { sendWelcomeEmbed } from "../../modules/guild/sendWelcomeEmbed";
+import { memberNicknameChange } from "../../modules/events/memberUpdate/memberNicknameChange";
+import { memberPassedScreening } from "../../modules/events/memberUpdate/memberPassedScreening";
+import { memberRolesChange } from "../../modules/events/memberUpdate/memberRolesChange";
 import { getSettings } from "../../modules/settings/getSettings";
 import { beccaErrorHandler } from "../../utils/beccaErrorHandler";
 
@@ -37,43 +38,7 @@ export const memberUpdate = async (
 
     // passes membership screening
     if (oldMember.pending && !newMember.pending) {
-      const welcomeText = (
-        serverSettings.custom_welcome || defaultServer.custom_welcome
-      )
-        .replace(/{@username}/gi, user.username)
-        .replace(/{@servername}/gi, guild.name);
-
-      if (serverSettings.welcome_style === "embed") {
-        const welcomeEmbed = new EmbedBuilder();
-        welcomeEmbed.setColor(Becca.colours.default);
-        welcomeEmbed.setTitle(t("events:member.join.title"));
-        welcomeEmbed.setDescription(welcomeText);
-        welcomeEmbed.setAuthor({
-          name: user.tag,
-          iconURL: user.displayAvatarURL(),
-        });
-        welcomeEmbed.setFooter({ text: `ID: ${user.id}` });
-        welcomeEmbed.setTimestamp();
-
-        await sendWelcomeEmbed(Becca, guild, "join", welcomeEmbed);
-      } else if (serverSettings.welcome_style === "text") {
-        const channel =
-          guild.channels.cache.get(serverSettings.welcome_channel) ||
-          (await guild.channels.fetch(serverSettings.welcome_channel));
-        if (channel && "send" in channel) {
-          await channel.send({
-            content: welcomeText,
-            allowedMentions: {},
-          });
-        }
-      }
-
-      if (serverSettings?.join_role) {
-        const joinRole = await guild.roles.fetch(serverSettings.join_role);
-        if (joinRole) {
-          await newMember.roles.add(joinRole);
-        }
-      }
+      await memberPassedScreening(Becca, newMember, serverSettings, t);
     }
 
     if (!serverSettings.member_events) {
@@ -104,55 +69,25 @@ export const memberUpdate = async (
     embed.setTimestamp();
 
     if (oldMember.nickname !== newMember.nickname) {
-      embed.setDescription(
-        t("events:member.update.description", {
-          key: "nickname",
-        })
+      await memberNicknameChange(
+        Becca,
+        oldMember,
+        newMember,
+        embed,
+        t,
+        logChannel
       );
-      embed.addFields([
-        {
-          name: t("events:member.update.old"),
-          value: oldMember.nickname || oldMember.user.username,
-          inline: true,
-        },
-        {
-          name: t("events:member.update.new"),
-          value: newMember.nickname || newMember.user.username,
-          inline: true,
-        },
-      ]);
-
-      await logChannel.send({ embeds: [embed] });
     }
 
     if (oldMember.roles.cache.size !== newMember.roles.cache.size) {
-      embed.setDescription(
-        t("events:member.update.description", {
-          key: "roles",
-        })
+      await memberRolesChange(
+        Becca,
+        oldMember,
+        newMember,
+        embed,
+        t,
+        logChannel
       );
-
-      const addedRoles = newMember.roles.cache
-        .filter((role) => !oldMember.roles.cache.has(role.id))
-        .map((role) => `<@&${role.id}>`);
-      const removedRoles = oldMember.roles.cache
-        .filter((role) => !newMember.roles.cache.has(role.id))
-        .map((role) => `<@&${role.id}>`);
-
-      embed.addFields([
-        {
-          name: t("events:member.update.old"),
-          value: removedRoles.join(", ") || "No roles removed.",
-          inline: true,
-        },
-        {
-          name: t("events:member.update.new"),
-          value: addedRoles.join(", ") || "No roles added.",
-          inline: true,
-        },
-      ]);
-
-      await logChannel.send({ embeds: [embed] });
     }
   } catch (err) {
     await beccaErrorHandler(
